+++ lib/bump/file_fetchers/python/pip.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+require "bump/file_fetchers/base"
+
+module Bump
+  module FileFetchers
+    module Python
+      class Pip < Bump::FileFetchers::Base
+        def self.required_files
+          %w(requirements.txt)
+        end
+      end
+    end
+  end
+end